@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>netcoreapp3.1</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
     <AssemblyName>Keyfactor.Extensions.Orchestrators.AKV</AssemblyName>
     <RootNamespace>Keyfactor.Extensions.Orchestrator.AzureKeyVault</RootNamespace>
 	  <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
@@ -18,30 +18,25 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-	  <None Remove="C:\Users\jvanwanzeele\.nuget\packages\keyfactor.extensions.pam.utilities\1.0.2\contentFiles\any\any\Keyfactor.Extensions.Pam.Config.exe.config" />
-	  <None Remove="C:\Users\jvanwanzeele\.nuget\packages\keyfactor.extensions.pam.utilities\1.0.2\contentFiles\any\any\Keyfactor.Extensions.Pam.Utilities.dll.config" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Azure.Core" Version="1.40.0" />
-		<PackageReference Include="Azure.Identity" Version="1.12.0" />
-		<PackageReference Include="Azure.ResourceManager" Version="1.12.0" />
-		<PackageReference Include="Azure.ResourceManager.KeyVault" Version="1.2.3" />
-		<PackageReference Include="Azure.ResourceManager.Resources" Version="1.7.3" />
-		<PackageReference Include="Azure.Security.KeyVault.Administration" Version="4.4.0" />
-		<PackageReference Include="Azure.Security.KeyVault.Certificates" Version="4.6.0" />
-		<PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.6.0" />
-		<PackageReference Include="Azure.Storage.Blobs" Version="12.20.0" />
-		<PackageReference Include="CSS.Common" Version="1.7.0" />
-		<PackageReference Include="Keyfactor.Common" Version="2.3.7" />
-		<PackageReference Include="Keyfactor.Extensions.Pam.Utilities" Version="1.0.2" />
-		<PackageReference Include="Keyfactor.Logging" Version="1.1.1" />
+		<PackageReference Include="Azure.Core" Version="1.44.1" />
+		<PackageReference Include="Azure.Identity" Version="1.13.1" />
+		<PackageReference Include="Azure.ResourceManager" Version="1.13.0" />
+		<PackageReference Include="Azure.ResourceManager.KeyVault" Version="1.3.0" />
+		<PackageReference Include="Azure.ResourceManager.Resources" Version="1.9.0" />
+		<PackageReference Include="Azure.Security.KeyVault.Administration" Version="4.5.0" />
+		<PackageReference Include="Azure.Security.KeyVault.Certificates" Version="4.7.0" />
+		<PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.7.0" />
+		<PackageReference Include="Azure.Storage.Blobs" Version="12.22.2" />
+		<PackageReference Include="Keyfactor.Common" Version="2.8.1" />
+		<PackageReference Include="Keyfactor.Logging" Version="1.2.0" />
 		<PackageReference Include="Keyfactor.Orchestrators.Common" Version="3.2.0" />
 		<PackageReference Include="Keyfactor.Orchestrators.IOrchestratorJobExtensions" Version="0.7.0" />
 		<PackageReference Include="Keyfactor.Platform.IPAMProvider" Version="1.0.0" />
 		<PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Identity.Client" Version="4.61.3" />
-		<PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="4.61.3" />
+		<PackageReference Include="Microsoft.Identity.Client" Version="4.66.1" />
+		<PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="4.66.1" />
+		<PackageReference Include="System.Linq" Version="4.3.0" />
+		<PackageReference Include="System.Linq.Async" Version="6.0.1" />
 	</ItemGroup>
 
 	<ItemGroup>
