@@ -17,9 +17,8 @@ using Microsoft.Extensions.Logging;
 namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
 {
     [Job(JobTypes.DISCOVERY)]
-    public class Discovery : AzureKeyVaultJob<Discovery>, IDiscoveryJobExtension    {
-        
-
+    public class Discovery : AzureKeyVaultJob<Discovery>, IDiscoveryJobExtension
+    {
         public Discovery(IPAMSecretResolver resolver)
         {
             PamSecretResolver = resolver;
@@ -28,7 +27,7 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
 
         public JobResult ProcessJob(DiscoveryJobConfiguration config, SubmitDiscoveryUpdate sdr)
         {
-            logger.LogDebug($"Begin Discovery...");
+            logger.LogDebug($"Begin Discovery job");
             InitializeStore(config);
 
             var complete = new JobResult() { JobHistoryId = config.JobHistoryId, Result = OrchestratorJobStatusJobResult.Failure };
@@ -47,30 +46,34 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
             }
 
             // if there are no warnings return vaults and status of success
-            if (warnings == null || !warnings.Any()) {
+            if (warnings == null || !warnings.Any())
+            {
                 logger.LogTrace("discovery completed with no warnings or errors.");
                 complete.Result = OrchestratorJobStatusJobResult.Success;
                 complete.FailureMessage = $"Discovery job completed successfully.  Found {keyVaults?.Count() ?? 0} KeyVaults.";
             }
 
             // if there are warnings, but vaults were found, return Vaults and status of warn
-            if (warnings?.Count() > 0 && keyVaults.Count() > 0) {
+            if (warnings?.Count() > 0 && keyVaults.Count() > 0)
+            {
                 logger.LogTrace("discovery completed with warnings.");
                 complete.Result = OrchestratorJobStatusJobResult.Warning;
                 complete.FailureMessage = $"Discovery job completed with errors.  Found {keyVaults?.Count() ?? 0} KeyVaults.\nThe following errors occurred: \n";
                 complete.FailureMessage = complete.FailureMessage + string.Join('\n', warnings);
-                if (complete.FailureMessage.Length > 4000) {
+                if (complete.FailureMessage.Length > 4000)
+                {
                     complete.FailureMessage = complete.FailureMessage.Substring(0, 3500) + "\n results truncated.  Please see the Orchestrator logs for more details.";
                 }
             }
 
             // if there are warnings and no vaults were found, return status of error
 
-            if (warnings?.Count() > 0 && keyVaults?.Count() == 0) {
+            if (warnings?.Count() > 0 && keyVaults?.Count() == 0)
+            {
                 logger.LogTrace("discovery completed with errors and no vaults found (failed).");
                 complete.Result = OrchestratorJobStatusJobResult.Failure;
                 complete.FailureMessage = $"Discovery job failed with the following errors: \n";
-                complete.FailureMessage = complete.FailureMessage + string.Join('\n', warnings);                
+                complete.FailureMessage = complete.FailureMessage + string.Join('\n', warnings);
             }
 
             // need to truncate failure message if it exceeds the max length of 4000
@@ -81,8 +84,8 @@ namespace Keyfactor.Extensions.Orchestrator.AzureKeyVault
             }
 
             sdr.Invoke(keyVaults);
-            
+
             return complete;
         }
-    }   
+    }
 }
